feat(hub-common): add searchCategoriesToCollections() helper function
affects: @esri/hub-common
